+++ src/ventaquil/OpenCLPi/BBPFormula.h
@@ -3,6 +3,6 @@
 
 #include <CL/cl.h>
 
-double BBPFormula(cl_uint s, cl_uint b, cl_uint m, cl_int *A, cl_uint k = 8);
+float BBPFormula(cl_uint s, cl_uint b, cl_uint m, cl_int *A, cl_uint k = 8);
 
 #endif //OPENCLPI_BBPFORMULA_H
